CLC-5068 - Moved Canvas::ProvideCourseSite.unique_job_id and .find to Canvas::BackgroundJob module
