+++ app/controllers/canvas_course_provision_controller.rb
@@ -57,7 +57,7 @@ class CanvasCourseProvisionController < ApplicationController
 
   # GET /api/academics/canvas/course_provision/status.json
   def job_status
-    course_provision_job = Canvas::ProvideCourseSite.find(params['job_id'])
+    course_provision_job = Canvas::BackgroundJob.find(params['job_id'])
     render json: course_provision_job.to_json and return if course_provision_job.class == Canvas::ProvideCourseSite
     render json: { job_id: params['job_id'], jobStatus: "jobNotFoundError", error: "Unable to find course management job" }.to_json
   end
